refactor(anyharness): move startup diagnostics to stderr driver
